+++ src/main/java/ru/job4j/ood/lsp/products/ControlQuality.java
@@ -1,7 +1,7 @@
-package ru.job4j.ood.lsp;
+package ru.job4j.ood.lsp.products;
 
-import ru.job4j.ood.lsp.model.Food;
-import ru.job4j.ood.lsp.store.Store;
+import ru.job4j.ood.lsp.products.model.Food;
+import ru.job4j.ood.lsp.products.store.Store;
 
 import java.time.LocalDate;
 import java.util.List;
